+++ src/DataMigrator.js
@@ -7,9 +7,9 @@ import PouchDB from './PouchDB';
 import clone from 'clone';
 
 import * as util from './util';
-import * as RxSchema from './RxSchema';
-import * as KeyCompressor from './KeyCompressor';
-import * as Crypter from './Crypter';
+import RxSchema from './RxSchema';
+import KeyCompressor from './KeyCompressor';
+import Crypter from './Crypter';
 
 class DataMigrator {
 
@@ -304,3 +304,7 @@ class OldCollection {
 export function create(newestCollection, migrationStrategies) {
     return new DataMigrator(newestCollection, migrationStrategies);
 }
+
+export default {
+    create
+};