+++ src/KeyCompressor.js
@@ -195,3 +195,7 @@ class KeyCompressor {
 export function create(schema) {
     return new KeyCompressor(schema);
 }
+
+export default {
+    create
+};