+++ src/RxBroadcastChannel.js
@@ -66,3 +66,8 @@ export function create(database, name) {
 
     return new RxBroadcastChannel(database, name);
 }
+
+export default {
+    create,
+    canIUse
+};