+++ src/Socket.js
@@ -1,7 +1,7 @@
-import * as RxCollection from './RxCollection';
-import * as RxChangeEvent from './RxChangeEvent';
-import * as RxBroadcastChannel from './RxBroadcastChannel';
 import * as util from './util';
+import RxCollection from './RxCollection';
+import RxChangeEvent from './RxChangeEvent';
+import RxBroadcastChannel from './RxBroadcastChannel';
 
 const EVENT_TTL = 5000; // after this age, events will be deleted
 const PULL_TIME = RxBroadcastChannel.canIUse() ? EVENT_TTL / 2 : 200;
@@ -169,4 +169,10 @@ export async function create(database) {
 export {
     EVENT_TTL as EVENT_TTL,
     PULL_TIME as PULL_TIME
+};
+
+export default {
+    create,
+    EVENT_TTL,
+    PULL_TIME
 };