+++ src/browserify.index.js
@@ -1,5 +1,5 @@
 import 'babel-polyfill';
-import * as RxDB from './index.js';
+import RxDB from './index.js';
 
 RxDB.plugin(require('pouchdb-adapter-idb'));
 RxDB.plugin(require('pouchdb-adapter-http'));