+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 5.x.x (coming soon)
+
+- Added default exports. Use `import RxDB from 'rxdb'` instead of `import * as RxDB from 'rxdb'`
+
 ## 5.2.1 (July 17, 2017)
 
 Quickfix because of new [pouchdb-import](https://github.com/pouchdb/pouchdb/issues/6603#issuecomment-315867346)